Collect registration/authentication logic in tom_common.accounts
The alternative was having this stuff scattered in modules all
over tom_common. Now it's in one place. (middleware and signals
remain in those modules).
