+++ docs/common/authentication.rst
@@ -157,7 +157,7 @@ point the setting at your version in ``settings.py`` — nothing in tom_base cha
   store them in ``signup(self, request, user)``; point django-allauth at it with
   ``ACCOUNT_SIGNUP_FORM_CLASS = 'custom_code.forms.MySignupForm'``.
 - **Behaviour** (who may sign up, which group new users join, notification emails, redirects): subclass
-  ``tom_common.adapters.TomAccountAdapter`` and set ``ACCOUNT_ADAPTER = 'custom_code.adapters.MyAccountAdapter'``.
+  ``tom_common.accounts.adapters.TomAccountAdapter`` and set ``ACCOUNT_ADAPTER = 'custom_code.adapters.MyAccountAdapter'``.
   See `Adapter <https://docs.allauth.org/en/latest/account/adapter.html>`_ for the available hooks.
 - **Templates**: ``account/signup.html``, ``account/signup_closed.html``, ``account/account_inactive.html`` and the
   email templates above can be overridden in your TOM's ``templates/`` directory.
@@ -222,10 +222,10 @@ Password rules
 Password rules are Django's ``AUTH_PASSWORD_VALIDATORS`` and apply everywhere a password is set (sign-up, change,
 reset, admin). TOM Toolkit adds two validators you can include:
 
-``tom_common.password_validation.CharacterClassValidator``
+``tom_common.accounts.password_validation.CharacterClassValidator``
     Requires at least one upper-case letter, one lower-case letter, one digit and one special character.
 
-``tom_common.password_validation.NotSameAsCurrentPasswordValidator``
+``tom_common.accounts.password_validation.NotSameAsCurrentPasswordValidator``
     Rejects a "new" password identical to the current one (useful together with expiry).
 
 For example, a policy of 12+ characters with all four character classes, different from the user's name and email::
@@ -234,8 +234,8 @@ For example, a policy of 12+ characters with all four character classes, differe
         {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator'},
         {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator', 'OPTIONS': {'min_length': 12}},
         {'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator'},
-        {'NAME': 'tom_common.password_validation.CharacterClassValidator'},
-        {'NAME': 'tom_common.password_validation.NotSameAsCurrentPasswordValidator'},
+        {'NAME': 'tom_common.accounts.password_validation.CharacterClassValidator'},
+        {'NAME': 'tom_common.accounts.password_validation.NotSameAsCurrentPasswordValidator'},
     ]
 
 You can write your own validator class in ``custom_code`` and list it the same way (see
@@ -322,12 +322,12 @@ API access::
   can only be (re)generated by the user themself, after re-authenticating; and the token is rejected while the
   account has an outstanding requirement (expired password, terms not accepted).
 
-Both are enforced by ``tom_common.api_auth.TomTokenAuthentication``, which replaces the plain REST framework
+Both are enforced by ``tom_common.accounts.api_auth.TomTokenAuthentication``, which replaces the plain REST framework
 ``TokenAuthentication`` in ``REST_FRAMEWORK``::
 
     REST_FRAMEWORK = {
         'DEFAULT_AUTHENTICATION_CLASSES': [
-            'tom_common.api_auth.TomTokenAuthentication',            # Authorization: Token <key>
+            'tom_common.accounts.api_auth.TomTokenAuthentication',            # Authorization: Token <key>
             'rest_framework.authentication.SessionAuthentication',   # logged-in browsers
             # 'rest_framework.authentication.BasicAuthentication',   # username/password per request — omit when
                                                                      # you require two-factor authentication
@@ -370,8 +370,8 @@ Forms
     replace individual forms; ``MFA_FORMS`` does the same for the 2FA forms. See
     `Forms <https://docs.allauth.org/en/latest/account/forms.html>`_.
 Adapters
-    ``ACCOUNT_ADAPTER`` (default ``tom_common.adapters.TomAccountAdapter``) controls registration, redirects,
-    email sending and the inactive-account response; ``MFA_ADAPTER`` (default ``tom_common.adapters.TomMFAAdapter``)
+    ``ACCOUNT_ADAPTER`` (default ``tom_common.accounts.adapters.TomAccountAdapter``) controls registration, redirects,
+    email sending and the inactive-account response; ``MFA_ADAPTER`` (default ``tom_common.accounts.adapters.TomMFAAdapter``)
     controls the issuer name, secret encryption and whether users may remove their authenticator. Subclass the
     TOM Toolkit adapter and override the hook you need.
 Password validators
@@ -380,10 +380,10 @@ Post-login requirements
     ``TOM_ACCOUNT_REQUIREMENTS`` is the ordered list of checks run for every logged-in request::
 
         TOM_ACCOUNT_REQUIREMENTS = [
-            'tom_common.account_requirements.terms_of_service_accepted',
-            'tom_common.account_requirements.mfa_enrolled',
-            'tom_common.account_requirements.password_not_expired',
-            'tom_common.account_requirements.required_fields_present',
+            'tom_common.accounts.requirements.terms_of_service_accepted',
+            'tom_common.accounts.requirements.mfa_enrolled',
+            'tom_common.accounts.requirements.password_not_expired',
+            'tom_common.accounts.requirements.required_fields_present',
         ]
 
     Each built-in check is a function whose companion ``TOM_*`` setting is its parameter: unconfigured means the
