+++ docs/common/customsettings.rst
@@ -171,10 +171,10 @@ Default:
 .. code-block:: python
 
    [
-       'tom_common.account_requirements.terms_of_service_accepted',
-       'tom_common.account_requirements.mfa_enrolled',
-       'tom_common.account_requirements.password_not_expired',
-       'tom_common.account_requirements.required_fields_present',
+       'tom_common.accounts.requirements.terms_of_service_accepted',
+       'tom_common.accounts.requirements.mfa_enrolled',
+       'tom_common.accounts.requirements.password_not_expired',
+       'tom_common.accounts.requirements.required_fields_present',
    ]
 
 Ordered list of checks applied to every request by a logged-in user. Each entry is the dotted path of a check
@@ -213,7 +213,7 @@ check may read its own ``settings.py`` value following the same pattern. See
 
 Default: None
 
-When set to a number of days, API tokens older than that are rejected by ``tom_common.api_auth.TomTokenAuthentication``
+When set to a number of days, API tokens older than that are rejected by ``tom_common.accounts.api_auth.TomTokenAuthentication``
 and users must regenerate their token from their profile edit page.
 
 `TOM_API_TOKEN_REQUIRES_MFA <#tom-api-token-requires-mfa>`__
@@ -223,7 +223,7 @@ Default: False
 
 When ``True``, API tokens are only accepted for users who have enabled two-factor authentication (and only tokens
 created after enrolment); the password-only ``/api/token-auth/`` endpoint is disabled and tokens can only be
-regenerated by their owner after re-authenticating. Requires ``tom_common.api_auth.TomTokenAuthentication`` in
+regenerated by their owner after re-authenticating. Requires ``tom_common.accounts.api_auth.TomTokenAuthentication`` in
 ``REST_FRAMEWORK['DEFAULT_AUTHENTICATION_CLASSES']``.
 
 `TOM_FACILITY_CLASSES <#tom-facility-classes>`__
